+++ src/components/post/ErrorAlert.tsx
